requirements.txt
asgiref==3.2.10
Django==3.0.8
django-environ==0.4.5
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1
mysqlclient==2.2.0
PyMySQL==1.1.0
# django-cors-headers==3.10.1
# djangorestframework==3.14.0
# djangorestframework-simplejwt==4.3.0
django-cors-headers
djangorestframework
djangorestframework-simplejwt